tests.image,.txt - separate text tests
